generator/requirements.txt
google-cloud-pubsub==2.18.0
faker==13.3.0
